cyber-pi-intel/pyproject.toml
[project]
name = "tqakb-v4"
version = "4.0.0"
description = "The Question Answer Knowledge Base - Next Generation Event-Driven Knowledge Management System"
authors = [{name = "David", email = "david@tqakb.ai"}]
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"

dependencies = [
    # Core Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    
    # Async Support
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "asyncio>=3.11.0",
    
    # Redis Stack
    "redis[hiredis]>=5.2.0",
    "redis-om>=0.3.3",
    
    # Neo4j Graph Database
    "neo4j>=5.25.0",
    "py2neo>=2021.2.4",
    
    # Weaviate Vector DB
    "weaviate-client>=4.9.0",
    
    # LLM & Embeddings
    "ollama>=0.4.0",
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-community>=0.3.0",
    
    # Data Processing
    "numpy>=2.1.0",
    "pandas>=2.2.0",
    "polars>=1.14.0",
    "arrow>=1.3.0",
    
    # Monitoring & Observability
    "prometheus-client>=0.21.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "structlog>=24.4.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
    "rich>=13.9.0",
    "typer>=0.15.0",
    "orjson>=3.10.0",
    "ujson>=5.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-benchmark>=5.0.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "ipython>=8.30.0",
    "ipdb>=0.13.0",
]

test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "faker>=33.0.0",
    "hypothesis>=6.119.0",
]

docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.27.0",
]

[build-system]
requires = ["setuptools>=75.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["backend", "backend.core", "backend.api", "backend.streaming"]

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = [".git", "__pycache__", "venv", ".venv", "build", "dist"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C90", "D"]
ignore = ["D100", "D101", "D102", "D103", "D104", "D107"]

[tool.black]
line-length = 120
target-version = ["py311"]
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=backend",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--asyncio-mode=auto",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_any_generics = false
ignore_missing_imports = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
strict_equality = true

[tool.coverage.run]
source = ["backend"]
omit = ["*/tests/*", "*/test_*.py", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]